feat: Add drizzle (#207)
* (feat) project wide drizzle config

This sets the drizzle instance for the project.

* small + regular drizzle svg's

added the drizzle svg's for demo.

* inserts db url to the .env file

add's example db url

* add's the db + driver chosen by user

they can choose, postgres, mysql and sqlite. Each has its own dependencies.

* sample to do schema

uses the user's chosen db and it's imports to create a todo db.

* initialize db instance

this creates a db instance to be used for queries throughout the project. It uses the user's chosen driver to set it up.

* create info file

added info file for drizzle db. let's user choose their db and driver.

* example demo route

used drizzle orm options to do the queries and test it out.

* feat: adding Drizzle

* fix: version bump

---------

Co-authored-by: Abdullahi Mohamed <126521894+WaryaWayne@users.noreply.github.com>